ensure fixtures with changed dates are normalized
if changing a fixture's date causese it to be outside it's GW window
it must be set as not normal, so that the gameweek's window isn't
expanded to the out-of-window changed date
