+++ server/crons/UpdateDate.go
@@ -1,6 +1,7 @@
 package crons
 
 import (
+	"fmt"
 	"time"
 
 	"github.com/kristo-og-logi/premKing/server/initializers"
@@ -17,7 +18,7 @@ func UpdateDate(dbFixture models.Fixture, jsonFixture models.SportmonksFixture) 
 	}
 
 	if jsonTime.Compare(dbFixture.MatchDate) != 0 {
-		// fmt.Printf("%s | from %s to %s\n", dbFixture.Name, dbFixture.MatchDate, jsonTime)
+		fmt.Printf("%s | from %s to %s\n", dbFixture.Name, dbFixture.MatchDate, jsonTime)
 		initializers.DB.Model(&dbFixture).Updates(models.Fixture{MatchDate: jsonTime})
 		updated = true
 	}
